+++ src/modal.js
@@ -1,4 +1,3 @@
-import apiCalls from './apiCalls';
 import comments from './comments';
 import counter from './counter';
 
@@ -27,12 +26,32 @@ class Modal {
 
     commentBtn.addEventListener('click', async () => {
       document.querySelector('.comments-cont').innerHTML = '';
-      apiCalls.mealPopulate(url);
+      this.mealPopulate(url);
       const info = await comments.getComments(dish);
       document.querySelector('.modalholder').id = id;
       comments.appendComment(info);
       const count = await counter.commentCounter(info);
       comments.appendCount(count);
+    });
+  }
+
+  mealRequest = async (url) => {
+    const returned = await fetch(url);
+    const meals = await returned.json();
+    return meals;
+  }
+
+  mealPopulate = async (url) => {
+    const modalWindow = document.querySelector('.modalholder');
+    this.mealRequest(url).then((data) => {
+      data.meals.forEach((element) => {
+        modalWindow.querySelector('.container__profilepic').src = element.strMealThumb;
+        modalWindow.querySelector('.item-name').innerText = element.strMeal;
+        modalWindow.querySelector('.area').innerText = `Area: ${element.strArea}`;
+        modalWindow.querySelector('.cat').innerText = `Category: ${element.strCategory}`;
+        modalWindow.querySelector('.demonstration').src = element.strYoutube;
+        modalWindow.querySelector('.ingre').innerText = `Main Ingredient: ${element.strIngredient1}`;
+      });
     });
   }
 }
